+++ programs/compressed-token/anchor/src/process_transfer.rs
@@ -247,7 +247,7 @@ pub fn create_output_compressed_accounts(
             owner: (*owner).to_anchor_pubkey().into(),
             amount: (*amount).into(),
             delegate: delegate.map(|delegate_pubkey| delegate_pubkey.into()),
-            state: AccountState::Initialized,
+            state: AccountState::Initialized as u8,
             tlv: None,
         };
         // TODO: remove serialization, just write bytes.
@@ -693,9 +693,9 @@ pub fn get_input_compressed_accounts_with_merkle_context_and_check_signer<const 
         };
         sum_lamports += compressed_account.lamports;
         let state = if IS_FROZEN {
-            AccountState::Frozen
+            AccountState::Frozen as u8
         } else {
-            AccountState::Initialized
+            AccountState::Initialized as u8
         };
         if input_token_data.tlv.is_some() {
             unimplemented!("Tlv is unimplemented.");
@@ -1185,7 +1185,7 @@ mod test {
                 mint: Pubkey::new_unique(),
                 owner: Pubkey::new_unique(),
                 delegate: None,
-                state: AccountState::Initialized,
+                state: AccountState::Initialized as u8,
                 amount: *i,
                 tlv: None,
             });
